Downgraded visibility of downlink factories.
